Convert Menu component to use SCSS modules
