+++ styles/globals.scss
@@ -11,7 +11,7 @@ body {
   padding: 0;
   margin: 0;
   font-family: "Rubik", sans-serif;
-  color: $text-primary;
+  color: $text-color-primary;
   overflow-x: hidden;
   @include font-size($body-font-sizes);
 }
